refactor: decompose stream_orchestrator.rs into focused module directory [OPE-291]
Split the monolithic stream_orchestrator.rs into a module directory:
- mod.rs: drive_stream orchestration function
- tests.rs: comprehensive test suite

Updated to incorporate current main changes (anyhow::Context error
handling, improved RecvError::Closed behavior).
